+++ cmd/fireeth/main.go
@@ -144,6 +144,7 @@ func Chain() *firecore.Chain[*pbeth.Block] {
 				parent.AddCommand(newFixAnyTypeCmd(zlog))
 				parent.AddCommand(newFixWithdrawalsCmd(zlog))
 				parent.AddCommand(newRemoveGasChangesCmd(zlog))
+				parent.AddCommand(newRemoveLargeKeccakPreimagesCmd(zlog))
 				parent.AddCommand(newPollRPCBlocksCmd(zlog))
 				parent.AddCommand(newPollerCmd(zlog, tracer))
 				parent.AddCommand(newOptimismPollerCmd(zlog, tracer))